backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.7.0
pyyaml>=6.0
openpyxl>=3.1.0
pypsa>=0.28.0
pandas>=2.2.0
numpy>=1.26.0
highspy>=1.7.0
